Move buffer factory and pdq buffer hasher to their own files
